Revert "Move schema resolving to separate helper class"
This reverts commit 75c59aad9977f0021001f823b36b791af954b662.
