+++ container-search/src/main/java/com/yahoo/prelude/cluster/ClusterSearcher.java
@@ -1,8 +1,8 @@
 // Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.prelude.cluster;
 
-import com.yahoo.component.ComponentId;
 import com.yahoo.component.annotation.Inject;
+import com.yahoo.component.ComponentId;
 import com.yahoo.component.chain.dependencies.After;
 import com.yahoo.component.provider.ComponentRegistry;
 import com.yahoo.container.QrSearchersConfig;
@@ -28,6 +28,9 @@ import com.yahoo.vespa.streamingvisitors.VdsStreamingSearcher;
 import com.yahoo.yolean.Exceptions;
 
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashSet;
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
@@ -53,7 +56,8 @@ public class ClusterSearcher extends Searcher {
 
     private final String searchClusterName;
 
-    private final SchemaResolver schemaResolver;
+    // The set of document types contained in this search cluster
+    private final Set<String> schemas;
 
     private final long maxQueryTimeout; // in milliseconds
     private final long maxQueryCacheTimeout; // in milliseconds
@@ -79,7 +83,7 @@ public class ClusterSearcher extends Searcher {
         searchClusterName = clusterConfig.clusterName();
         QrSearchersConfig.Searchcluster searchClusterConfig = getSearchClusterConfigFromClusterName(qrsConfig, searchClusterName);
         this.globalPhaseRanker = searchClusterConfig.globalphase() ? globalPhaseRanker : null;
-        this.schemaResolver = new SchemaResolver(documentDbConfig);
+        schemas = new LinkedHashSet<>();
 
         maxQueryTimeout = ParameterParser.asMilliSeconds(clusterConfig.maxQueryTimeout(), DEFAULT_MAX_QUERY_TIMEOUT);
         maxQueryCacheTimeout = ParameterParser.asMilliSeconds(clusterConfig.maxQueryCacheTimeout(), DEFAULT_MAX_QUERY_CACHE_TIMEOUT);
@@ -87,6 +91,9 @@ public class ClusterSearcher extends Searcher {
         SummaryParameters docSumParams = new SummaryParameters(qrsConfig
                 .com().yahoo().prelude().fastsearch().FastSearcher().docsum()
                 .defaultclass());
+
+        for (DocumentdbInfoConfig.Documentdb docDb : documentDbConfig.documentdb())
+            schemas.add(docDb.name());
 
         String uniqueServerId = UUID.randomUUID().toString();
         if (searchClusterConfig.indexingmode() == STREAMING) {
@@ -149,7 +156,7 @@ public class ClusterSearcher extends Searcher {
 
     /** Do not use, for internal testing purposes only. **/
     ClusterSearcher(Set<String> schemas, VespaBackEndSearcher searcher, Executor executor) {
-        this.schemaResolver = new SchemaResolver(schemas);
+        this.schemas = schemas;
         searchClusterName = "testScenario";
         maxQueryTimeout = DEFAULT_MAX_QUERY_TIMEOUT;
         maxQueryCacheTimeout = DEFAULT_MAX_QUERY_CACHE_TIMEOUT;
@@ -221,9 +228,8 @@ public class ClusterSearcher extends Searcher {
     }
 
     private Result doSearch(Searcher searcher, Query query, Execution execution) {
-        var schemas = schemaResolver.resolve(query, execution);
         if (schemas.size() > 1) {
-            return searchMultipleDocumentTypes(searcher, query, execution, schemas);
+            return searchMultipleDocumentTypes(searcher, query, execution);
         } else {
             String docType = schemas.iterator().next();
             query.getModel().setRestrict(docType);
@@ -259,7 +265,8 @@ public class ClusterSearcher extends Searcher {
         }
     }
 
-    private Result searchMultipleDocumentTypes(Searcher searcher, Query query, Execution execution, Set<String> schemas) {
+    private Result searchMultipleDocumentTypes(Searcher searcher, Query query, Execution execution) {
+        Set<String> schemas = resolveSchemas(query, execution.context().getIndexFacts());
         List<Query> queries = createQueries(query, schemas);
         if (queries.size() == 1) {
             return perSchemaSearch(searcher, queries.get(0), execution);
@@ -293,7 +300,25 @@ public class ClusterSearcher extends Searcher {
     }
 
     Set<String> resolveSchemas(Query query, IndexFacts indexFacts) {
-        return schemaResolver.resolve(query, indexFacts);
+        Set<String> restrict = query.getModel().getRestrict();
+        if (restrict == null || restrict.isEmpty()) {
+            Set<String> sources = query.getModel().getSources();
+            return (sources == null || sources.isEmpty())
+                    ? schemas
+                    : new HashSet<>(indexFacts.newSession(sources, Set.of(), schemas).documentTypes());
+        } else {
+            return filterValidDocumentTypes(restrict);
+        }
+    }
+
+    private Set<String> filterValidDocumentTypes(Collection<String> restrict) {
+        Set<String> retval = new LinkedHashSet<>();
+        for (String docType : restrict) {
+            if (docType != null && schemas.contains(docType)) {
+                retval.add(docType);
+            }
+        }
+        return retval;
     }
 
     private List<Query> createQueries(Query query, Set<String> docTypes) {
